requirements.txt
pillow
opencv-python
snekchek
git+https://github.com/squeaky-pl/japronto
